requirements.txt
opencv-python
mediapipe
numpy
pynput
python-uinput
PyYAML
